Todo puede ser un componente qwq
