+++ src/components/menu/MenuItemForm.js
@@ -11,12 +11,33 @@ import { Toast } from 'primereact/toast';
 import { Tooltip } from 'primereact/tooltip';
 import { ProgressBar } from 'primereact/progressbar';
 
-const MenuItemForm = ({ isVisible, onHide, isEditMode, formValues, setFormValues, handleSave }) => {
+const MenuItemForm = ({ isVisible, onHide, isEditMode, formValues, setFormValues, addMenuItem, loadMenuItems}) => {
     const toast = useRef(null);
     const imageUploaded = useRef(null);
     const [totalSize, setTotalSize] = useState(0);
     const chooseOptions = { icon: 'pi pi-fw pi-images', iconOnly: true, className: 'custom-choose-btn p-button-rounded p-button-outlined' };
     const cancelOptions = { icon: 'pi pi-fw pi-times', iconOnly: true, className: 'custom-cancel-btn p-button-danger p-button-rounded p-button-outlined' };
+
+    const handleSave = async () => {
+        if (isEditMode) {
+            console.log("Editando producto:", formValues);
+            // Aquí se podría llamar a una función de edición si se necesita
+        } else {    
+            try {
+                await addMenuItem(formValues);
+                console.log("Producto agregado exitosamente:", formValues);
+                toast.current.show({ severity: 'success', summary: 'Éxito', detail: 'Producto agregado', life: 3000 });
+    
+                // Recargar la lista de ítems para reflejar el nuevo producto
+                await loadMenuItems();
+                setTotalSize(0);
+                onHide(); // Cierra el modal después de guardar exitosamente
+            } catch (error) {
+                console.error('Error adding item:', error);
+                toast.current.show({ severity: 'error', summary: 'Error', detail: 'No se pudo agregar el producto', life: 3000 });
+            }
+        }
+    };
 
     // Función para manejar la selección de un archivo y calcular su tamaño
     const onTemplateSelect = (e) => {
@@ -33,7 +54,6 @@ const MenuItemForm = ({ isVisible, onHide, isEditMode, formValues, setFormValues
 
         setTotalSize(_totalSize); // Actualiza el estado del tamaño total con el nuevo valor
         //toast.current.show({ severity: 'info', summary: 'Éxito', detail: 'Archivo Subido' }); // Muestra una notificación de éxito
-
     };
     
     // Función para manejar la carga del archivo y mostrar una notificación de éxito
@@ -48,7 +68,6 @@ const MenuItemForm = ({ isVisible, onHide, isEditMode, formValues, setFormValues
         setTotalSize(_totalSize);
         toast.current.show({ severity: 'info', summary: 'Éxito', detail: 'Archivo Subido' }); // Muestra una notificación de éxito
     };
-
 
     const onTemplateClear = () => {
         setTotalSize(0);
